# musics/api/swagger.yaml
basePath: /
definitions:
...
  models.Genre:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  models.Music:
    properties:
      albumId:
        type: string
      artistId:
        type: string
      genreId:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
info:
  contact:
    email: lucas.denis@etu.uca.fr
    name: Lucas Denis.
  description: Spotifaux's renowned API.
  title: Spotifaux
  version: 1.0.0
paths:
...
  /genres:
    get:
      description: Get genres.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "500":
          description: Something went wrong
      summary: Get genres.
      tags:
      - genres
    post:
      description: Create a Genre.
      parameters:
      - description: Genre object
        in: body
        name: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "500":
          description: Something went wrong
      summary: Create a Genre.
      tags:
      - genres
...
  /musics:
    get:
      description: Get musics.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Music'
            type: array
        "500":
          description: Something went wrong
      summary: Get musics.
      tags:
      - musics
    post:
      description: Create a music.
      parameters:
      - description: Music object
        in: body
        name: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Music'
            type: array
        "500":
          description: Something went wrong
      summary: Create a music.
      tags:
      - musics
  /musics/{id}:
    delete:
      description: Delete a music.
      parameters:
      - description: Music UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Delete a music.
      tags:
      - musics
    get:
      description: Get a music.
      parameters:
      - description: Music UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Music'
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Get a music.
      tags:
      - musics
    put:
      description: Update a music.
      parameters:
      - description: Music UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      - description: Music object
        in: body
        name: body
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Update a music.
      tags:
      - musics
schemes:
- http
swagger: "2.0"